requirements.txt
aiosmtpd==1.4.4.post2
pycryptodome
colorama
bcrypt
jxdb